requirements.txt
aiogram==2.15
aiohttp==3.7.4.post0
aioredis==2.0.0
astroid==2.8.2
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.6
click==8.0.2
enum34==1.1.10
hiredis==2.0.0
idna==3.2
idna-ssl==1.1.0
isort==5.9.3
itsdangerous==2.0.1
Jinja2==3.0.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
multidict==5.2.0
pycodestyle==2.7.0
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
rope==0.20.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typed-ast==1.4.3
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
yarl==1.7.0
